Home magnitude_bin_index with the bin geometry types
Bin-index arithmetic operates on MagnitudeCountBounds and belongs in models
beside the geometry types, not in the action-metadata registry. This decouples
ranking math from the aggregate registry import surface.
